Add files via upload
As part of the Codsoft programming assignment, the objective was to design and implement a professional-grade calculator using Python. The task encompassed creating a versatile calculator capable of performing basic arithmetic operations such as addition, subtraction, multiplication, and division. The solution needed to adhere to industry best practices, ensuring clarity, conciseness, and correctness in the code.